feat: implement list tables and add crud example

+++ src/rebox-storage/src/database/row/column/mod.rs
@@ -34,8 +34,8 @@ impl TableColumn {
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub enum ColumnValue {
     Bool(bool),
-    Integer(i32),
-    Natural(u32),
+    Integer(i64),
+    Natural(u64),
     Text(String),
 }
